SSE includes this file in its build and complained about ambiguous call, added cast to double.

+++ CreatorCode/SolimarLicensing/common/ProcessorUtils.h
@@ -120,7 +120,7 @@ private:
 	 * 
 	 * Calculates log base 2.
 	 */
-	double log2(double n) { return log(n) / log(2); }
+	double log2(double n) { return log(n) / log((double)2); }
 
 	/*
 	 * ParseAPICID()
